bAmazonSeeds_1.sql: 1. Created bamazon_db database and products table. 2. Inserted mock data for products.
bAmazonCustomer.js: 1. Load NPM packages: mysql, inquirer and clear. 2. Set up connection to MYSQL. 3. Connect to bamazon_db. 4. Display info of products for sale. 5. Ask user to choose a product that they would like to buy and specify the quantity. 6. Validate if the quantity is available. 7. If it is, update the stock remaining and display the order's total. 8. If it isnt', let the user know that the quantity exceeds the current stock.